source/api/serializers.py
from rest_framework import serializers

from webapp.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ('id', 'title', 'description', 'status', 'type', 'project', 'is_deleted', 'created_at', 'updated_at')
        read_only_fields = ('id', 'type','project')

    def create(self, validated_data):
        return Task.objects.create(**validated_data)

    def update(self, instance, validated_data):
        for key, value in validated_data.items():
            setattr(instance, key, value)
        instance.save()
        return instance